+++ clash-systemverilog/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`clash-systemverilog`](http://hackage.haskell.org/package/clash-systemverilog) package
 
+## 0.6.6
+* New features:
+  * Support `clash-prelude` 0.10.7
+
 ## 0.6.5 *January 29th 2016*
 * New features:
   * Support clash-lib-0.6.9